:sparkles: [김채영] double형으로 휴가 개수 지급 & 메인에 휴가 지급 출력

+++ src/main/java/com/fiveLink/linkOffice/mapper/VacationMapper.java
@@ -36,7 +36,7 @@ public interface VacationMapper {
 
     int checkTypeName( Map<String, Object> map);
 
-    int userVacationCount(Long memberNo);
+    double userVacationCount(Long memberNo);
 
     String memberHireDate(Long memberNo);
 }
